pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "capacity-planner"
version = "0.1.0"
description = "Network capacity deployment optimizer built with Pyomo and Dash."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "dash>=2.18,<3.0",
  "numpy>=1.26,<3.0",
  "pandas>=2.2,<3.0",
  "plotly>=5.24,<6.0",
  "pyomo>=6.8,<7.0",
  "pyyaml>=6.0,<7.0",
  "highspy>=1.8,<2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
